# Store only type pointer for Node/Type self-ref, not full interface