--- a/Herebyfile.mjs
+++ b/Herebyfile.mjs
@@ -27,6 +27,7 @@ const { values: options } = parseArgs({
         race: { type: "boolean" },
         fix: { type: "boolean" },
         noembed: { type: "boolean" },
+        unsafe: { type: "boolean" },
         debug: { type: "boolean" },
         concurrentTestPrograms: { type: "boolean" },
     },
@@ -40,6 +41,7 @@ const { values: options } = parseArgs({
 
 const defaultGoBuildTags = [
     ...(options.noembed ? ["noembed"] : []),
+    ...(options.unsafe ? ["tsunsafe"] : []),
 ];
 
 /**