--- a/internal/checker/checker.go
+++ b/internal/checker/checker.go
@@ -22937,17 +22937,6 @@ func isUnaryTupleTypeNode(node *ast.Node) bool {
 	return ast.IsTupleTypeNode(node) && len(node.AsTupleTypeNode().Elements.Nodes) == 1
 }
 
-func (c *Checker) newType(flags TypeFlags, objectFlags ObjectFlags, data TypeData) *Type {
-	c.TypeCount++
-	t := data.AsType()
-	t.flags = flags
-	t.objectFlags = objectFlags &^ (ObjectFlagsCouldContainTypeVariablesComputed | ObjectFlagsCouldContainTypeVariables | ObjectFlagsMembersResolved)
-	t.id = TypeId(c.TypeCount)
-	t.checker = c
-	t.data = data
-	return t
-}
-
 func (c *Checker) newIntrinsicType(flags TypeFlags, intrinsicName string) *Type {
 	return c.newIntrinsicTypeEx(flags, intrinsicName, ObjectFlagsNone)
 }
@@ -22995,7 +22984,7 @@ func (c *Checker) newUniqueESSymbolType(symbol *ast.Symbol, name string) *Type {
 }
 
 func (c *Checker) newObjectType(objectFlags ObjectFlags, symbol *ast.Symbol) *Type {
-	var data TypeData
+	var data typeData
 	switch {
 	case objectFlags&ObjectFlagsClassOrInterface != 0:
 		data = &InterfaceType{}