@jakebailey 2025-03-15
I rebased this to simplify it, but I actually negated the entire point since the struct embedding seems to add some padding, oops.